use apt-get remove; remove/purge all packages at once
Removing packages with dpkg one-by-one does not respect dependencies
and may fail e.g. when Pre-Depends are involved.  Therefore use
  apt-get remove   on all packages in one run to reorder them properly.
apt-get purge can't be used because it fails on unknown packages (i.e. not
in the archive and not in config-files-remaining state). See #637853.
Also apt-get in lenny is not able to properly purge packages in
config-files-remaining state.
But we can purge all packages in one dpkg call.

The remove_or_purge() method is replaced by two new methods
remove_packages() and purge_packages().

Signed-off-by: Andreas Beckmann <debian@abeckmann.de>