read output remaining after command terminated
There may be output generated between reading a chunk (or at the
very beginning) and the time Popen.poll() returns not None for
the first time - this output was previously missed. So read again
from the Popen.stdout file after the process has terminated.
Limit the amount that is read, it is possible that a buggy program
forks into the background, keeps the stdout/stderr file descriptors
and produces infinite output but is not killed by Popen.terminate()
and friends ... (These remaining programs will be terminated by
the check for running processes.)

Hopefully fixes the spurious missing output from dpkg-divert --list
and other commands.

Signed-off-by: Andreas Beckmann <debian@abeckmann.de>