check for network issues before daily report
Run detect_network_issues as part of the daily processing.
We are not interested in having these issues as unclassified
problems in the daily report nor having piuparts-analyze try to
map them to existing reports.
Just list them as waiting-to-be-tested.

Keep the cronjob, so that these are rescheduled twice per day.

Signed-off-by: Andreas Beckmann <debian@abeckmann.de>