a post_remove script to aid debugging purge errors
Add 'set -x' in the postrm scripts of $PIUPARTS_OBJECTS before
running 'postrm purge'.

This is helpful for packages failing in the purge phase to find
the actual failing command.

This is only intended for debugging purge failures and should not
be enabled by default.  But you can easily rerun the piuparts
command manually and add another instance of --scriptsdir to
include this script.

Signed-off-by: Andreas Beckmann <debian@abeckmann.de>