only newest version is a candidate
Packages files may contain several versions of the same package and
are in unspecified ordering. This may happen especially for arch:all
packages.
Discard all duplicate package entries but the *newest* version while
reading Packages files. (Up to now the *last* entry was used, but
this is not neccessarily the installation candidate, i.e. the newest.)

Signed-off-by: Andreas Beckmann <debian@abeckmann.de>