do not care about old versions of untestable packages
If an old log is in untestable/ (usually because of some dependency
issue) do not let this block testing of a newer package version.
The new package may fix the issue. Just test the new package like if
it were in any other state. Do not wait for reschedule_untestable.

Signed-off-by: Andreas Beckmann <debian@abeckmann.de>