t-a: write BTS links to a .bug file per failed .log
Write a <package>_<version>.bug file with a link ot the BTS whenever
a log is moved to affected/ or bugged/ (or can't because the version
in the bug(s) found did not match).
This can be picked up by the detect_well_known_errors script and
inserted into the generated *.tpl.

Signed-off-by: Andreas Beckmann <debian@abeckmann.de>