Add exception for python-support broken symlinks.
python-support creates symlinks upon install under /usr/lib/pythonx.y,
for all supported versions of python, whether installed or not.
If the version is not installed, the symlink is broken (#635493).
If a supported python version subsequently uninstalled, the symlink
is left dangling (#385775). Fixing this involves changes across
packages that are unlikely to happen for a deprecated package.

DP Appendix B.2
http://www.debian.org/doc/packaging-manuals/python-policy/ap-packaging_tools.html

Signed-off-by: Andreas Beckmann <anbe@debian.org>