piuparts-report.py - Avoid sorting passed files by block count.
And refactor passed vs. failed sort handling.

There was still a case where a full accounting of rdep metrics was
called for - the sort of successfully-tested in piuparts-report.
Avoid the sort call to block_count if we know it is not going to
affect the sort.

With this change, piuparts-reports now runs at 35 seconds per section.