conf.py: Add get_std_distro()
Try to determine the right 'standard' distribution for the section,
either 'oldstable', 'stable', 'testing', 'unstable', or 'experimental',
based on the settings from get_distro() and get_distros().
Returns 'unknown' if undetermined.

The tests used for development are included as test-config.py

This adds a "Depends" for python-distro-info to piuparts-common, as a
new component in conf.py.