Remove oldstable and older from p.d.o summary
There is a desire in the DDPO integration to add value to the Piuparts
column by showing more detail in the failures, and to winnow out chaff
in the reporting.

The details on how to do this are not fully developed. In the
meantime, as a first, interim, step, configure p.d.o to remove summary
reporting for section tests involving oldstable and older.