dependencyparser.py: loosen the multiarch spec
Per the Multi Arch spec, a package dependency may have an ":any" appended to
the package name in a Depends-type spec. The dependencyparser module detects
and strips any occurance of ":any".

https://wiki.ubuntu.com/MultiarchSpec#Extended_semantics_of_per-architecture_package_relationships

Per #748426, lsb-core has an ":i386" multiarch modifier on a package name,
causing the dependencyparser to crash in piuparts-report, disabling all p.d.o
reporting.

http://lists.alioth.debian.org/pipermail/piuparts-devel/2014-June/005468.html

Leaving the proper lsb-core behavior to another venue, this change modifies
dependencyparser to be lenient about the ":" modifiers it accepts and strips.