Revert "Don't add results from outdated binaries to summary.json"
This reverts commit 62a6b3c438f42480ce3a08996ed4994de1561591.

The patch was added to avoid incorporating summary.json reporting
results for binary packages which are no longer part of the source
package. It stopped Piuparts from blocking the source package from
testing migration in this case.

It has a bug related to "upgrade test" sections, which cause
a large number of packages to incorrectly be reported in the "waiting"
state.

Revert, and rely on ftp-masters to decruft.

https://alioth-lists.debian.net/pipermail/piuparts-devel/2019-March/008422.html