piupartslib/__init__: fix byte stream decoding
When trying to decode a byte sequence into a string, it is possible that
the sequence ends in the middle of a multibyte character. In this case
the decoding fails.

This hack assumes that the input byte sequence is always encoded in
UTF-8, and that the original data stream never ends in the middle of a
multibyte character.

Signed-off-by: Holger Levsen <holger@layer-acht.org>