Make a valid pkgsummary entry for an "invalid" pkg
pkgsummary was assuming that package entries would never be made with an
"invalid" state. In that case, the entry would end up with the default
url "invalid url", which never should be the result. When an actual
package with an "invalid" state is entered, the eventual global summary
creation will fail on the bad url. Fix by always replacing a default
entry.

Test included.

Signed-off-by: Holger Levsen <holger@layer-acht.org>